# src/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Internship Task API
  description: Scalable REST API with Authentication & Role-Based Access
  version: 1.0.0
servers:
  - url: http://localhost:5000/api/v1
    description: Local Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /tasks:
    get:
      summary: Get all tasks for the logged-in user
      tags:
        - Tasks
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
    post:
      summary: Create a new task
      tags:
        - Tasks
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - description
              properties:
                title:
                  type: string
                description:
                  type: string
                status:
                  type: string
      responses:
        '201':
          description: Created
...